refactor critterfactory
